+++ models/document_contributor.go
@@ -3,5 +3,5 @@ package models
 type DocumentContributor struct {
 	UserID     string `gorm:"type:uuid;primary_key;default:uuid_generate_v4()"`
 	DocumentID string `gorm:"type:uuid;primary_key;default:uuid_generate_v4()"`
-	UpdateTime int64  `gorm:"type:bigint"`
+	UpdateTime int32  `gorm:"type:bigint"`
 }
